Fixed an inifinte loop in ping_server.pde
When client moves out of range previous code would cause the server to get stuck in an infinite loop trying to send. When the client were to move back in range the server would not respond because it was in the loop.

I have added a small timeout code which prevents this, at the cost of performance.